Move polar plotting functions
Move from plotting/polar_map.py to lib/plotting_functions.py

This will keep with the consistency of plotting functions in one file
